requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
